Spring Project: all tests passed

+++ src/main/java/com/example/service/MessageService.java
@@ -32,6 +32,12 @@ public class MessageService {
         return mr.findAll();
     }
 
+    //return all messages for a specific user (postedBy references accountId in account table)
+    public List<Message> getAllMessagesByUser(int postedBy){
+        return mr.findAllByPostedBy(postedBy);
+    }
+
+
     //get one message by its messageId
     public Message getMessageById(int id){
         Optional<Message> optMessage = mr.findById(id);
@@ -47,18 +53,34 @@ public class MessageService {
     //public Message getMessageby  (){}
 
     //delete message, maybe unnecessary
-    public void deleteMessage(int id){
-        mr.deleteById(id);
+    public Message deleteMessage(int id){
+        Message temp = getMessageById(id);
+        if(temp == null){
+            return null;
+        }
+        else{
+            mr.deleteById(id);
+            return temp;
+        }
     }
 
     //update message text, returns changed message or null if non-existant
+    //checks for blank message text and length over 255 characters (again, directions inconsistent)
+    //doesnt necessarily need to return a whole Message object for this particular use case
+    //but I think it's more universal this way
     public Message updateMessage(int id, Message newMessage){
+        //logic check newMessage, from above (short circuits everything after)
+        if(newMessage.getMessageText().length() > 255 || newMessage.getMessageText().isEmpty()){
+            return null;
+        }
+        //otherwise
         Optional<Message> optMessage = mr.findById(id);
         if(optMessage.isPresent()){
             Message temp = optMessage.get();
             temp.setMessageText(newMessage.getMessageText());
             mr.save(temp);
             return temp;
+
         }
         else{
             return null;
